Move ktlint configuration creation method to Configurations file.
